requirements.txt
opencv-python
Pillow
gradio-image-prompter
sentencepiece
torch>=2.0.0
flash-attn>=2.3.3
